update submodule, breast limit multiplier

+++ app/src/main/cpp/GakumasLocalify/config/Config.cpp
@@ -29,7 +29,7 @@ namespace GakumasLocal::Config {
     int lodQualityLevel = 4;
 
     bool enableBreastParam = false;
-    int bUseLimit = 1;
+    float bUseLimit = 1.0f;
     float bDamping = 0.33f;
     float bStiffness = 0.08f;
     float bSpring = 1.0f;
